+++ NEWS.md
@@ -1,3 +1,9 @@
+# pharmaversesdtm (development version)
+
+## New Features
+
+-  Updated QSTESTCD in `qs_ophtha` to be consistent with CDISC terminology. Also reduced the number of records for ease of use. (#172)
+
 # pharmaversesdtm 1.3.1
 
 ## Documentation